tools/migrate/requirements.txt
# Python 3.12 — Access (.accdb) -> Postgres migration tool (ADR-010).
#
# pyodbc reads the Access file via the mdbtools ODBC driver. macOS:
#   brew install unixodbc mdbtools && \
#   echo $'[MDBTools]\nDriver=/opt/homebrew/lib/libmdbodbc.so' >> ~/.odbcinst.ini
# Ubuntu 24.04:
#   apt install unixodbc-dev mdbtools libmdbodbc1
#
# psycopg connects as a role that has been GRANTed platform_admin_role
# (BYPASSRLS) — the tool issues `SET ROLE platform_admin_role` at session
# start so RLS doesn't block cross-clinic inserts.
pyodbc==5.2.0
psycopg[binary]==3.2.3
PyYAML==6.0.2
